mass refactor now compiles
